refactor: move onSpanEndCallback logic out of batch processor
