+++ packages/core/tests/span-factory.test.ts
@@ -146,7 +146,7 @@ describe('SpanFactory', () => {
         const clock = new IncrementingClock('1970-01-01T00:00:00.000Z')
         const sampler = new Sampler(0.5)
         const delivery = { send: jest.fn() }
-        const processor = { add: (span: SpanEnded) => delivery.send(spanToJson(span, clock)), runCallbacks: async () => true }
+        const processor = { add: (span: SpanEnded) => delivery.send(spanToJson(span, clock)) }
         const backgroundingListener = new ControllableBackgroundingListener()
         const spanFactory = new SpanFactory(
           processor,
@@ -169,7 +169,7 @@ describe('SpanFactory', () => {
         const clock = new IncrementingClock('1970-01-01T00:00:00.000Z')
         const sampler = new Sampler(0.5)
         const delivery = { send: jest.fn() }
-        const processor = { add: (span: SpanEnded) => delivery.send(spanToJson(span, clock)), runCallbacks: async () => true }
+        const processor = { add: (span: SpanEnded) => delivery.send(spanToJson(span, clock)) }
         const backgroundingListener = new ControllableBackgroundingListener()
         const spanFactory = new SpanFactory(
           processor,
@@ -193,7 +193,7 @@ describe('SpanFactory', () => {
         const clock = new IncrementingClock('1970-01-01T00:00:00.000Z')
         const sampler = new Sampler(0.5)
         const delivery = { send: jest.fn() }
-        const processor = { add: (span: SpanEnded) => delivery.send(spanToJson(span, clock)), runCallbacks: async () => true }
+        const processor = { add: (span: SpanEnded) => delivery.send(spanToJson(span, clock)) }
         const backgroundingListener = new ControllableBackgroundingListener()
         const spanFactory = new SpanFactory(
           processor,
@@ -217,7 +217,7 @@ describe('SpanFactory', () => {
         const clock = new IncrementingClock('1970-01-01T00:00:00.000Z')
         const sampler = new Sampler(0.5)
         const delivery = { send: jest.fn() }
-        const processor = { add: (span: SpanEnded) => delivery.send(spanToJson(span, clock)), runCallbacks: async () => true }
+        const processor = { add: (span: SpanEnded) => delivery.send(spanToJson(span, clock)) }
         const backgroundingListener = new ControllableBackgroundingListener()
         const spanFactory = new SpanFactory(
           processor,
